Modified   Project/Networking/Binaries/Resources/InitialData_Default.json Modified   Project/Networking/Server/Data/GeneralData.cs Modified   Project/Networking/Server/Data/TableData.cs Modified   Project/Networking/Server/FriendlyRoom.cs Modified   Project/Networking/Server/OneByBotRoom.cs Modified   Project/Networking/Server/OneByOneRoom.cs Modified   Project/Networking/Server/Room.cs

+++ Project/Networking/Server/OneByOneRoom.cs
@@ -19,7 +19,7 @@ namespace Networking.Server
 			get { return null; }
 		}
 
-		public OneByOneRoom(Application Application, int TableID, float TurnTime) :
+		public OneByOneRoom(Application Application, int TableID, int TurnTime) :
 			base(Application, TableID, TurnTime)
 		{
 		}
